Backdev -> B0-ReadingList로 최신화

+++ ReadingLogBackend/src/main/java/com/example/demo/readingrecord/ReadingRecord.java
@@ -28,7 +28,7 @@ public class ReadingRecord {
     private long num;
 
     @Column(name = "book_id")
-    private long bookId;
+    private Integer bookId;
 
     @Column(name = "user_id")
     private Integer userId;
